gateway/pyproject.toml
[project]
name = "fsx-gateway"
version = "0.1.0"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "jinja2==3.1.6",
  "psycopg2-binary==2.9.9",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
